extract into new file
